RCB-123 First pass at reducing 429s
