+++ lib/Api.js
@@ -107,7 +107,7 @@ Api.prototype.rosette = function(endpoint, callback) {
             // send parameters to the specified endpoint
             e.getResults(api.parameters, api.userKey, api.serviceURL, function(err, res) {
                 if (err) {
-                    return callback(err);
+                    throw err;
                 } else {
                     return callback(null, res);
                 }
