+++ lib/categories.js
@@ -108,7 +108,8 @@ categories.prototype.getResults = function(parameters, userKey, serviceURL, call
                     if (res.statusCode === 200) {
                         return callback(err, JSON.parse(result.toString()));
                     } else if (res.statusCode != 200) {
-                        return callback(err, JSON.parse(result.toString()));
+                        return callback(new RosetteException(res.statusCode, result.toString()));
+                        //return callback(err, JSON.parse(result.toString()));
                     }
                 });
             });
