+++ lib/checkVersion.js
@@ -20,13 +20,7 @@ var zlib = require("zlib");
 
 var rosetteConstants = require("./rosetteConstants");
 var RosetteException = require("./rosetteExceptions");
-
-/**
- * Compatible server version.
- *
- * @type string
- */
-var BINDING_VERSION = "1.0";
+var rosetteRequest = require("./rosetteRequest");
 
 /**
  * @class
@@ -45,68 +39,11 @@ function checkVersion() {
  * @param {function} callback - Callback function to be exectuted after the function to which it is passed is complete
  */
 checkVersion.prototype.check = function(userKey, serviceURL, callback) {
+    var req = new rosetteRequest();
     // configure URL
-    var urlParts = URL.parse(serviceURL + "info?clientVersion=" + BINDING_VERSION);
-    var protocol = https;
-    if (urlParts.protocol === "http:") {
-        protocol = http;
-    }
+    var urlParts = URL.parse(serviceURL + "info?clientVersion=" + req.bindingVersion());
 
-    var headers = {
-        "accept": "application/json",
-        "accept-encoding": "gzip",
-        "content-type": "application/json",
-        "user-agent": "rosetteapinode/" + BINDING_VERSION
-    }
-    headers["X-RosetteAPI-Key"] = userKey;
-
-    var result = new Buffer("");
-
-    var options = {
-        hostname: urlParts.hostname,
-        path: urlParts.path,
-        method: 'POST',
-        headers: headers,
-        agent: false 
-    };
-
-    if (urlParts.port) {
-        options.port = urlParts.port;
-    }
-
-    // execute the http/https request
-    var req = protocol.request(options, function(res) {
-        res.on("data", function(docs) {
-            result = Buffer.concat([result, docs]);
-        });
-
-        res.on("end", function(err) {
-            if (err) {
-                return callback(err)
-            }
-
-            if (res.headers["content-encoding"] === "gzip") {
-                result = zlib.gunzipSync(result);
-            }
-
-            result = JSON.parse(result);
-            result.headers = res.headers;
-            result = JSON.stringify(result);
-
-            if (res.statusCode === 200) {
-                return callback(err, JSON.parse(result.toString()));
-            } else if (res.statusCode != 200) {
-                return callback(err, JSON.parse(result.toString()));
-            }
-        });
-    });
-
-    req.on("error", function(e) {
-        console.log(e)
-        return callback(e);
-    });
-
-    req.end();
+    req.makeRequest(userKey, urlParts, null, callback);
 
 };
