+++ lib/language.js
@@ -20,13 +20,7 @@ var zlib = require("zlib");
 
 var rosetteConstants = require("./rosetteConstants");
 var RosetteException = require("./rosetteExceptions");
-
-/**
- * Compatible server version.
- *
- * @type string
- */
-var BINDING_VERSION = "1.0";
+var rosetteRequest = require("./rosetteRequest");
 
 /**
  * @class
@@ -59,74 +53,8 @@ language.prototype.getResults = function(parameters, userKey, serviceURL, callba
         } else {
             // configure URL
             var urlParts = URL.parse(serviceURL + "language");
-            var protocol = https;
-            if (urlParts.protocol === "http:") {
-                protocol = http;
-            }
-
-            var contentType = "application/json";
-            if (parameters.loadParams().multipart === true) {
-                contentType = "multipart/mixed";
-            }
-
-            var headers = {
-                "accept": "application/json",
-                "accept-encoding": "gzip",
-                "content-type": "application/json",
-                "user-agent": "rosetteapinode/" + BINDING_VERSION
-            }
-            headers["X-RosetteAPI-Key"] = userKey;
-
-            var result = new Buffer("");
-
-            var options = {
-                hostname: urlParts.hostname,
-                path: urlParts.path,
-                method: 'POST',
-                headers: headers,
-                agent: false
-            };
-
-            if (urlParts.port) {
-                options.port = urlParts.port;
-            }
-
-            // execute the http/https request
-            var req = protocol.request(options, function(res) {
-                res.on("data", function(docs) {
-                    result = Buffer.concat([result, docs]);
-                });
-
-                res.on("end", function(err) {
-                    if (err) {
-                        return callback(err)
-                    }
-                    // will need for other endpoints
-                    if (res.headers["content-encoding"] === "gzip") {
-                        result = zlib.gunzipSync(result);
-                    }
-
-                    result = JSON.parse(result);
-                    result.headers = res.headers;
-                    result = JSON.stringify(result);
-
-                    if (res.statusCode === 200) {
-                        return callback(err, JSON.parse(result.toString()));
-                    } else if (res.statusCode != 200) {
-                        return callback(err, JSON.parse(result.toString()));
-                    }
-                });
-            });
-
-            // error catching for request
-            req.on("error", function(e) {
-                return callback(e);
-            });
-
-            req.write(JSON.stringify(parameters.loadParams()));
-
-            req.end();
-
+            var req = new rosetteRequest();
+            req.makeRequest(userKey, urlParts, parameters, callback);
         }
     }
